+++ packages/mongo/mongo_driver.js
@@ -829,7 +829,7 @@ MongoConnection.prototype.findOne = function (collection_name, selector,
 
 // We'll actually design an index API later. For now, we just pass through to
 // Mongo's, but make it synchronous.
-MongoConnection.prototype._createIndex = function (collectionName, index,
+MongoConnection.prototype.createIndex = function (collectionName, index,
                                                    options) {
   var self = this;
 
@@ -841,7 +841,7 @@ MongoConnection.prototype._createIndex = function (collectionName, index,
   future.wait();
 };
 
-MongoConnection.prototype._ensureIndex = MongoConnection.prototype._createIndex;
+MongoConnection.prototype._ensureIndex = MongoConnection.prototype.createIndex;
 
 MongoConnection.prototype._dropIndex = function (collectionName, index) {
   var self = this;